Generate password support. Nomad support.
When some service starts (i.e. in test environment) it needs all configuration from consul and vault. So we generate password if it not exists.
In production we generate password and can overwrite it manually later.

When this image starts under some manager (i.e. nomad), it write config and exit. Then manager restarts it. So we need to start and stay started.

+ dockerfile refactoring, using alpine:edge image (to use String::Random module), manually install consul-cli, support for empty parameters (passwords), support for empty consul token.